agent/requirements.txt
fastapi
uvicorn[standard]
langgraph
langchain-core
langchain-openai
sentence-transformers
numpy
httpx
python-dotenv
